moves cra-todo to react-todo-list
