+++ examples/react-todo-list/.gitignore
@@ -10,6 +10,7 @@
 
 # production
 /build
+/.idea
 
 # misc
 .DS_Store
@@ -17,9 +18,8 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.eslintcache
 
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
-.vercel
