+++ examples/react-todo-list/.prettierrc
