+++ examples/react-todo-list/README.md
@@ -1,59 +1,81 @@
-## Collaborative Realtime Task Lists
+# Todo example using Supabase
 
-![Alt Text](https://dev-to-uploads.s3.amazonaws.com/i/n48hnd9zl78a0uq68xak.png)
+- Frontend:
+    - [create-react-app](https://reactjs.org/docs/create-a-new-react-app.html). React Framework for building single-page applications.
+    - [Tailwind](https://tailwindcss.com/) for styling and layout.
+    - [Supabase.js](https://supabase.io/docs/library/getting-started) for user management and realtime data syncing.
+- Backend:
+    - [app.supabase.io/](https://app.supabase.io/): hosted Postgres database with restful API for usage with Supabase.js.
 
-## Tutorial
 
-[https://dev.to/awalias/howto-build-collaborative-realtime-task-lists-in-react-4k52](https://dev.to/awalias/howto-build-collaborative-realtime-task-lists-in-react-4k52)
+## Demo
 
-## Warning
+-not added-
 
-This demo doesn't provide any form of client-side authentication so please don't use it in
-production unless you're ok with all your tasks being publicly available
+## Deploy with Vercel
 
-## Required env vars
+-not added-
 
-Make a `.env` in this folder with the following, 
-you can get these values from your project dashboard at [app.supabase.io](app.supabase.io):
+## Build from scratch
 
-```bash
-REACT_APP_SUPABASE_URL=<your-supabase-url>
-REACT_APP_SUPABASE_KEY=<your-supabase-key>
+### 1. Create new project
+
+Sign up to Supabase - [https://app.supabase.io](https://app.supabase.io) and create a new project. Wait for your database to start.
+
+### 2. Run "Todo List" Quickstart
+
+Once your database has started, run the "Todo List" quickstart.
+
+### 3. Get the URL and Key
+
+Go to the Project Settings (the cog icon), open the API tab, and find your API URL and `anon` key, you'll need these in the next step.
+
+The `anon` key is your client-side API key. It allows "anonymous access" to your database, until the user has logged in. Once they have logged in, the keys will switch to the user's own login token. This enables row level security for your data. Read more about this [below](#postgres-row-level-security).
+
+![image](https://user-images.githubusercontent.com/10214025/88916245-528c2680-d298-11ea-8a71-708f93e1ce4f.png)
+
+**_NOTE_**: The `service_role` key has full access to your data, bypassing any security policies. These keys have to be kept secret and are meant to be used in server environments and never on a client or browser.
+
+
+You will be asked for a `REACT_APP_SUPABASE_URL` and `REACT_APP_SUPABASE_KEY`. Use the API URL and `anon` key from [step 3](#3.-get-the-url-and-key).
+
+## Supabase details
+
+### Postgres Row level security
+
+This project uses very high-level Authorization using Postgres' Role Level Security.
+When you start a Postgres database on Supabase, we populate it with an `auth` schema, and some helper functions.
+When a user logs in, they are issued a JWT with the role `authenticated` and thier UUID.
+We can use these details to provide fine-grained control over what each user can and cannot do.
+
+This is a trimmed-down schema, with the policies:
+
+```sql
+create table todos (
+  id bigint generated by default as identity primary key,
+  user_id uuid references auth.users not null,
+  task text check (char_length(task) > 3),
+  is_complete boolean default false,
+  inserted_at timestamp with time zone default timezone('utc'::text, now()) not null
+);
+
+alter table todos enable row level security;
+
+create policy "Individuals can create todos." on todos for
+    insert with check (auth.uid() = user_id);
+
+create policy "Individuals can view their own todos. " on todos for
+    select using (auth.uid() = user_id);
+
+create policy "Individuals can update their own todos." on todos for
+    update using (auth.uid() = user_id);
+
+create policy "Individuals can delete their own todos." on todos for
+    delete using (auth.uid() = user_id);
 ```
 
-## Available Scripts
+## Authors
 
-In the project directory, you can run:
+- [Supabase](https://supabase.io)
 
-### `yarn start`
-
-Runs the app in the development mode.<br />
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
-
-The page will reload if you make edits.<br />
-You will also see any lint errors in the console.
-
-### `yarn test`
-
-Launches the test runner in the interactive watch mode.<br />
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
-
-### `yarn build`
-
-Builds the app for production to the `build` folder.<br />
-It correctly bundles React in production mode and optimizes the build for the best performance.
-
-The build is minified and the filenames include the hashes.<br />
-Your app is ready to be deployed!
-
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
-
-### `yarn eject`
-
-**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
-
-If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
-
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
-
-You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
+Supabase is open source, we'd love for you to follow along and get involved at https://github.com/supabase/supabase
