+++ examples/react-todo-list/public/index.html
@@ -1,5 +1,4 @@
 <!DOCTYPE html>
-<base href="/" />
 <html lang="en">
   <head>
     <meta charset="utf-8" />
@@ -8,9 +7,8 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Create and share task lists amongst friends and colleagues"
+      content="CRA Supabase TodoList Example"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
@@ -25,7 +23,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Supabase Task List</title>
+    <title>Supabase TodoList Example</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
