+++ examples/react-todo-list/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "CRA Supabase TodoList Example",
+  "name": "Create React App Supabase TodoList Example",
   "icons": [
     {
       "src": "favicon.ico",
