+++ examples/react-todo-list/src/index.css
@@ -1,37 +1,13 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
 body {
   margin: 0;
+  padding: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-color: rgb(250, 217, 246);
-}
-
-h1 {
-  margin-bottom: 0;
-}
-
-.container {
-  margin: 1em;
-}
-
-.section {
-  margin: 1em;
-}
-
-.build {
-  width: 600px;
-}
-
-.build-img {
-  max-width: 100%;
-  max-height: 100%;
-  display: block;
-}
-
-@media screen and (max-width : 600px) {
-  .build {
-    width: 300px;
-  }
 }
