+++ examples/react-todo-list/tailwind.config.js
